+++ src/test/java/domain/lotto/LotteryMatchTest.java
@@ -1,0 +1,17 @@
+package domain.lotto;
+
+import org.junit.jupiter.api.DisplayName;
+import org.junit.jupiter.api.Test;
+
+import static org.assertj.core.api.Assertions.assertThat;
+
+public class LotteryMatchTest {
+    @DisplayName("matchCount와 bonusMatch를 입력받아 LotteryMatch를 생성한다")
+    @Test
+    void matchTest() {
+        LotteryMatch lotteryMatch = new LotteryMatch(3, true);
+
+        assertThat(lotteryMatch).isEqualTo(new LotteryMatch(3, true));
+    }
+
+}
